+++ src/index.ts
@@ -1,3 +1,7 @@
-export { glob } from "./glob/glob.js";
+export { glob, globSync, globStream } from "./glob/glob.js";
 export type { GlobOptions, GlobEntry } from "./types/index.js";
 export { GlobError, PatternSyntaxError, WalkError } from "./utils/errors.js";
+export { matchPath, matchGlob } from "./glob/match.js";
+export type { MatchOptions } from "./glob/match.js";
+export { parsePattern } from "./glob/pattern.js";
+export { expandPatterns } from "./glob/expand.js";