+++ src/utils/fs.ts
@@ -1,4 +1,4 @@
-import { promises as fsp } from "node:fs";
+import { promises as fsp, readdirSync as fsReaddirSync, statSync as fsStatSync } from "node:fs";
 import type { Dirent } from "node:fs";
 
 export async function readdir(dir: string): Promise<Dirent[]> {
@@ -7,4 +7,12 @@ export async function readdir(dir: string): Promise<Dirent[]> {
 
 export async function stat(path: string) {
   return fsp.stat(path);
+}
+
+export function readdirSync(dir: string): Dirent[] {
+  return fsReaddirSync(dir, { withFileTypes: true });
+}
+
+export function statSync(path: string) {
+  return fsStatSync(path);
 }